Clean up app-utils based on self-review
- Update readme
- Delete AppRootBootstrap (unused)
