+++ packages/app-utils/README.md
@@ -1,8 +1,6 @@
 # @deephaven/app-utils
 
-A library with some utility functions used by Deephaven to load and manage plugins.
-
-For more details on using Deephaven JS Plugins, see the [deephaven-js-plugins repository](https://github.com/deephaven/deephaven-js-plugins).
+A library with some utility functions used by Deephaven applications.
 
 ## Install
